GH-39 Restructure ai-configuration directory and add toggle command
