+++ packages/ai-core/src/browser/ai-configuration/agent-configuration-widget.tsx
