+++ packages/ai-core/src/browser/ai-configuration/ai-configuration-widget.tsx
@@ -17,9 +17,9 @@
 import { BaseWidget, BoxLayout, codicon, DockPanel, WidgetManager } from '@theia/core/lib/browser';
 import { TheiaDockPanel } from '@theia/core/lib/browser/shell/theia-dock-panel';
 import { inject, injectable, postConstruct } from '@theia/core/shared/inversify';
-import '../../src/browser/style/index.css';
-import { AIAgentConfigurationWidget } from './agent-configuration/agent-configuration-widget';
-import { AIVariableConfigurationWidget } from './agent-configuration/variable-configuration-widget';
+import '../../../src/browser/style/index.css';
+import { AIAgentConfigurationWidget } from './agent-configuration-widget';
+import { AIVariableConfigurationWidget } from './variable-configuration-widget';
 
 @injectable()
 export class AIConfigurationContainerWidget extends BaseWidget {
