+++ packages/ai-core/src/browser/ai-configuration/language-model-renderer.tsx
