+++ packages/ai-core/src/browser/ai-configuration/template-settings-renderer.tsx
