+++ packages/ai-core/src/browser/ai-configuration/variable-configuration-widget.tsx
